requirements.txt
pytz
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-paginate==2022.1.8
Flask-Migrate==4.0.5
gunicorn==21.2.0
Flask-Babel==4.0.0
psycopg2-binary==2.9.7